% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binDiff.R
\name{binDiff}
\alias{binDiff}
\title{Compute interactions correlations for each bin between 2 matrices}
\usage{
binDiff(
  mutated.mat,
  wildtype.mat,
  bin.width,
  start = NULL,
  stop = NULL,
  max.distance = 2e+06,
  method = "Corr",
  seqname = "1",
  log2 = TRUE
)
}
\arguments{
\item{mutated.mat, wildtype.mat}{mutated (query) or wildtype (control, subject) matrices as \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{start, stop}{Area in bp of the chromosome to be computed. Default is NULL to use the entire chromosome (i.e. entire matrix).}

\item{max.distance}{Maximum distance between interactions. Default is 2e6 (i.e. 2Mb): to keep only interactions between 2 bins with a distance smaller than 2Mb}

\item{method}{Default is "Corr" for pearson correlation. Use "FC" to compute the mean fold change between interactions.}

\item{seqname}{chromosome names as character, default = "1".}

\item{log2}{if TRUE, log2 the interaction values of the matrices before computing the correlation. Default is TRUE.}
}
\value{
integer
}
\description{
For each bin, compute the correlation between the mutated and wildtype matrices.
}
\examples{
# to do

}
