% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicpro2matrix.R
\name{hicpro2matrix}
\alias{hicpro2matrix}
\title{Import matrix from HiC-Pro files}
\usage{
hicpro2matrix(matrix.path, index.path, chr, verbose = TRUE)
}
\arguments{
\item{matrix.path}{The path to the HiC-Pro .matrix file.}

\item{index.path}{The path to the HiC-Pro .bed index file.}

\item{chr}{The selected chromosome.}

\item{verbose}{Logical. Whether or not to print messages. Default = \code{TRUE}.}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix.
}
\description{
\code{hicpro2matrix} imports the interaction matrix for one chromosome from HiC-Pro output files (.matrix and .bed, see details) as a \code{dgCMatrix} (upper triangular and sparse Matrix).
By default, it assumes the input file contains only the upper triangular part of the matrix.
}
\details{
HiC-Pro is a tool for Hi-C data processing (https://github.com/nservant/HiC-Pro).
The \code{.matrix} file contains 3 columns: bin i, bin j, and the number of interactions.
The \code{.bed} file (index) contains 4 columns: chromosome, start, end, and bin ID.
\code{hicpro2matrix} filters the interactions to keep only intra-chromosomal interactions for the specified chromosome.
The resulting matrix is upper triangular.
}
