% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interMATplot.R
\name{interMATplot}
\alias{interMATplot}
\title{Plot intra-chromosomal interaction matrix}
\usage{
interMATplot(
  matrix,
  bin.width,
  start_y = NULL,
  stop_y = NULL,
  start_x = NULL,
  stop_x = NULL,
  log2 = T,
  scale.colors = "H",
  scale.limits = NULL,
  tad.x.tri = NULL,
  tad.y.tri = NULL,
  loop.bedpe = NULL,
  tad.chr_x = NULL,
  tad.chr_y = NULL,
  annotations.color = "red",
  tad.x.line = NULL,
  tad.y.line = NULL,
  tad.line.col = NULL,
  line.colors = c("red", "blue")
)
}
\arguments{
\item{matrix}{non-symmetrical \code{dgCMatrix} or \code{matrix} object with inter-chromosomal interactions.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{start_y, stop_y}{Region of interest in base pair on y-axis. Default is NULL to plot all the chromosome.}

\item{start_x, stop_x}{Region of interest in base pair on x-axis. Default is NULL to plot all the chromosome.}

\item{log2}{logical. Use the log2 of the matrix values. Default is \code{TRUE}.}

\item{scale.colors}{A character string indicating the color map option to use. Eight colors palettes are available from \code{viridis} package. Two supplementary palettes \code{"OE"} and  \code{"OE2"} (blue to red and purple to green respectively) are made for data centered on 0 (e.g log2 of observed/expected matrix). Default is \code{"H"}:
\itemize{
\item \code{"magma"} (or \code{"A"}),
\item \code{"inferno"} (or \code{"B"}),
\item \code{"plasma"} (or \code{"C"}),
\item \code{"viridis"} (or \code{"D"}),
\item \code{"cividis"} (or \code{"E"}),
\item \code{"rocket"} (or \code{"F"}),
\item \code{"mako"} (or \code{"G"}),
\item \code{"turbo"} (or \code{"H"}),
\item \code{"ObsExp"} (or \code{"OE"}),
\item \code{"ObsExp2"} (or \code{"OE2"}).
}}

\item{scale.limits}{Use to set limits on the scale. Default is NULL to use all values. E.g if use scale.limits = c(0, 10): Values < 0 will be fix to 0 and values > 10 will be fix to 10.}

\item{tad.y.tri, tad.x.tri}{\code{data.frame}, \code{GRanges} or the bed files path with the TAD to plot as triangle in the lower (x) or upper (y) part of the diagonal. Default is \code{NULL}.}

\item{loop.bedpe}{\code{data.frame} or bedpe files path to plot on the matrix. Six columns table (chr_y, start_y, end_y, chr_x, start_x, end_x) that gives areas between 2 chromosomes. Default is \code{NULL}}

\item{annotations.color}{Color for loop and triangular annotations. Default is \code{"red"}.}

\item{tad.y.line, tad.x.line}{\code{data.frame}, \code{GRanges} or the bed files path with the TAD to plot as line in the x or y axis of the matrix. Default is \code{NULL}.}

\item{tad.line.col}{Column number of \code{tad.y.line} and \code{tad.x.line} files that contain factors used to color lines. Default is \code{NULL}.}

\item{line.colors}{Colors for \code{tad.x.line} and \code{tad.y.line}. Default is \code{c("red", "blue")}.}

\item{tad.chr.y, tad.chr.x}{Chromosome name to filter annotations (domains and loop). Default is \code{NULL}}
}
\value{
\code{ggplot}
}
\description{
\code{interMATplot} allow to plot non-symmetrical matrix with 2 types of annotations:
\itemize{
\item domains (e.g. TADs or compartments): plot as triangles and/or lines on the x or/and y axis of the matrix.
\item interactions between domains/bins (e.g. loops): plot as squares or rectangles.
}
}
\details{
The matrix input must be a \code{Matrix} or a \code{matrix} object with interactions from 2 chromosome (i.e. inter-chromosomal interaction).
All domains (TADs or compartments) are bed files (3 columns: chr, start and end) and can be R object (\code{dataframe} or \code{GRanges}) or the path of the files.
For \code{tad.lines}, another column can be used to specify different classes of domains (e.g compartment A or B). To use those domain classes, specify the column number (from \code{tad.x.line} and \code{tad.y.line} inputs) with \code{tad.line.col} parameter and a custom set of colors with \code{line.colors} parameter.
Loop are stored in bedpe files (6 columns: chr_y, start_y, end_y, chr_x, start_x and end_x) and can be a \code{dataframe} object or the path of the file.
Chromosome domains and loops can be filter using \code{tad.chr.y} and \code{tad.chr.y} parameter.
}
\examples{

# create a matrix 500x1000 bins
matrix <- matrix(rnorm(500000, mean = 50, sd = 10) + c(rep(0, 250), seq(0, 50, length.out = 500), rep(0, 250)), 500, 1000)

interMATplot(matrix,
    bin.width = 100e3)

}
