% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matDI.R
\name{matDI}
\alias{matDI}
\title{Calculate Directionality Index (DI) on matrices}
\usage{
matDI(
  matrix.lst,
  bin.width,
  DI_distance = 5e+05,
  start = NULL,
  stop = NULL,
  seqname = "NA"
)
}
\arguments{
\item{matrix.lst}{List of \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Integer. The matrix bin.width in base pairs (e.g., 100000).}

\item{DI_distance}{Integer. The observation window distance (default is 500kb). It is recommended to use a distance between 400kb to 1mb according to bin.width (e.g. 500kb for bin.width = 16kb, 1mb for bin.width = 128kb).}

\item{start}{Integer. Start coordinate of the region of interest in bp (optional).}

\item{stop}{Integer. Stop coordinate of the region of interest in bp (optional).}

\item{seqname}{Character. Chromosome name, default = "NA".}
}
\value{
A list of \code{GRanges} objects with DI scores.
}
\description{
This function calculates the DI (stabilized with log1p).
}
\examples{
matDI(matrix.lst = list(HCT116 = mat_HCT116_chr19_50kb),
  bin.width = 50e3, DI_distance = 500e3,
  start = 5e6, stop = 15e6, seqname = "chr19")

}
