% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matIS.R
\name{matIS}
\alias{matIS}
\title{Calculate Insulation Score (IS) on matrices}
\usage{
matIS(
  matrix.lst,
  bin.width,
  square_size = 3e+05,
  start = NULL,
  stop = NULL,
  seqname = "NA"
)
}
\arguments{
\item{matrix.lst}{List of \code{dgCMatrix} or \code{matrix} objects for a single chromosome.}

\item{bin.width}{Integer. The matrix bin width in base pairs (e.g., 100000).}

\item{square_size}{Integer. The size of the sliding square window (default: 300kb).}

\item{start}{Integer. Start coordinate of the region of interest (optional).}

\item{stop}{Integer. Stop coordinate of the region of interest (optional).}

\item{seqname}{Character. Chromosome name, default = "NA".}
}
\value{
A list of \code{GRanges} objects with IS scores.
}
\description{
This function calculates the Insulation Score stabilized with log1p.
}
\details{
The Insulation Score (IS) is calculated for each bin as the mean of the interaction counts in a square window centered on the bin.
The final IS values are normalized the mean IS across the analyzed region. Therefore considering the entire chromosome for better normalization is recommended.
}
\examples{
matIS(matrix.lst = list(HCT116 = mat_HCT116_chr19_50kb),
  bin.width = 50e3, square_size = 300e3,
  start = 5e6, stop = 15e6, seqname = "chr19")

}
